refactoring icons
