+++ src/icons/WordPress.jsx
