+++ src/templates/portafolio.js
@@ -4,7 +4,7 @@ import Img from "gatsby-image"
 import Markdown from "@danestves/react-markdown"
 
 import Layout from "../components/layout"
-import { CategoryIcon, TechnologyIcon, LinkIcon } from "../components"
+import { CategoryIcon, TechnologyIcon, LinkIcon } from "../icons"
 
 export default ({ data }) => {
   const portfolio = data.strapiPortfolios
